Replace old influx codec
