+++ tremor-pipeline/Cargo.toml
@@ -14,6 +14,7 @@ serde_yaml = "0.8"
 petgraph = "0.5"
 error-chain="0.12"
 tremor-script = { path = "../tremor-script" }
+tremor-influx = { path = "../tremor-influx" }
 lazy_static = "1"
 window = { git = "https://github.com/wayfair-tremor/window.git" }
 indexmap = { version = "1", features=["serde-1"] }
