+++ tremor-script/Cargo.toml
@@ -9,8 +9,6 @@ build = "build.rs"
 readme = "README.md"
 documentation = "https://docs.rs/tremor-script"
 
-
-
 [package.metadata.docs.rs]
 default-target = "x86_64-unknown-linux-gnu"
 
@@ -57,6 +55,7 @@ sketches-ddsketch = "0.1.1"
 xz2 = "0.1"
 url = "2"
 libmath = "0.2"
+tremor-influx = { path = "../tremor-influx" }
 
 [build-dependencies]
 lalrpop = "0.17"
